Implement generator resource, add egui system boilerplate
