+++ src/server/terrain/systems.rs
@@ -1,20 +1,9 @@
-use bevy::ui::Node;
-use bevy_flair::style::components::NodeStyleSheet;
-
 use crate::prelude::*;
 
 pub fn setup_world_system(
     mut chunk_manager: ResMut<ChunkManager>,
-    params: Res<TerrainGeneratorParams>,
+    generator: Res<terrain_resources::Generator>,
 ) {
-    let generator = terrain_util::generator::Generator::new(
-        0,
-        TerrainGeneratorParams {
-            height_params: params.height_params,
-            density_params: params.density_params,
-        },
-    );
-
     let render_distance = Vec3::new(12.0, 2.0, 12.0);
 
     info!("Generating chunks");
@@ -29,23 +18,15 @@ pub fn setup_world_system(
     chunk_manager.insert_chunks(chunks);
 }
 
-// visualizer
+pub use visualizer::*;
 
-struct NoiseImageNode {}
+mod visualizer {
 
-pub fn setup_visualizer_system(
-    mut commands: Commands,
-    mut asset_server: ResMut<AssetServer>,
-) {
-    commands.spawn((
-            Node::default(),
-            Name::new("visualizer"),
-            NodeStyleSheet::new(asset_server.load("visualizer.css")),
-    )).with_children(|parent| {
-        parent.spawn((
-                Node::default(),
-                ImageNode::new(),
-                Name::new("noise_image"),
-        )
-    })
+    use bevy_inspector_egui::{bevy_egui::EguiContexts, egui};
+
+    pub fn render_visualizer_system(mut contexts: EguiContexts) {
+        egui::Window::new("Hello").show(contexts.ctx_mut(), |ui| {
+            ui.label("world");
+        });
+    }
 }
