+++ src/server/terrain/util/mod.rs
@@ -2,4 +2,3 @@ pub mod blocks;
 pub mod generator;
 
 pub use blocks::*;
-pub use generator::*;
